finishing lab1 and LSH

+++ src/CompareSignatures.scala
@@ -1,7 +1,9 @@
 object CompareSignatures {
 
- //  estimates similarity of two integer vectors – minhash signatures – as a fraction of components, in which they agree.
-	def compare_sig(sig1: Array[Double] , sig2: Array[Double], minHashLength:Int ): Double = {
+ 
+  
+  //  estimates similarity of two integer vectors – minhash signatures – as a fraction of components, in which they agree.
+	def compare_sig(sig1: Array[Int] , sig2: Array[Int], minHashLength:Int ): Double = {
 			var agreements = 0 
 					for (i<- 0 to minHashLength-1  ) {
 						if (sig1(i) == sig2(i))
@@ -12,8 +14,8 @@ object CompareSignatures {
 
 
 	def main(args: Array[String]): Unit = {
-			var s:Array[Double] =Array(1,3,3,7)
-			var t:Array[Double] = Array(77,20,3,-5)
+			var s:Array[Int] =Array(1,3,3,7)
+			var t:Array[Int] = Array(77,20,3,-5)
 			println(compare_sig(s,t , s.size));
 	}
